feat: Make the code more secure with env vars
Moved the models and tokenizers under artifacts directory, decoupled the downloading of the artifacts process from the server file and the method is imported from other file from the utils directory
